perf: :zap: exclude specific folders from the input via config file
This will avoid bundle generation error when non includeble json files are found in the input folder. As os.walk wil walk down all the subfolders of the input directory. The previous hardcoding of folders within the generateBundle is now removed.

The conf.json file should be updated according to the conf-example.json to avoid any breaks when running pyfhir